+++ app/src/main/java/com/nytimes/android/external/register/di/ActivityComponent.kt
@@ -2,14 +2,15 @@ package com.nytimes.android.external.register.di
 
 import com.nytimes.android.external.register.BuyActivity
 import com.nytimes.android.external.register.MainActivity
+import com.nytimes.android.external.register.MainFragment
 import com.nytimes.android.external.register.SettingsFragment
-
 import dagger.Subcomponent
 
 @Subcomponent(modules = [ActivityModule::class])
 @ScopeActivity
 interface ActivityComponent {
     fun inject(activity: MainActivity)
+    fun inject(activity: MainFragment)
     fun inject(activity: BuyActivity)
     fun inject(activity: SettingsFragment)
 }
