+++ app/src/main/res/layout/content_legal.xml
@@ -9,4 +9,4 @@
     app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
     tools:listitem="@layout/item_legal"
-    tools:showIn="@layout/activity_main" />
+    tools:showIn="@layout/fragment_main" />
